+++ src/statespace/dart/SO3Joint.cpp
@@ -1,9 +1,4 @@
-#include <aikido/statespace/dart/SO3Joint.hpp>
-
-// TODO: This will not be necessary once we switch to using
-// BallJoint::convertToRotation instead of calling dart::math::logMap directly.
-// See the comment below for more information.
-#include <dart/math/Geometry.hpp>
+#include "aikido/statespace/dart/SO3Joint.hpp"
 
 using ::dart::dynamics::BallJoint;
 
@@ -12,29 +7,27 @@ namespace statespace {
 namespace dart {
 
 //==============================================================================
-SO3Joint::SO3Joint(::dart::dynamics::BallJoint* _joint)
-  : SO3(), JointStateSpace(_joint)
+SO3Joint::SO3Joint(const ::dart::dynamics::BallJoint* joint)
+  : SO3(), JointStateSpace(joint)
 {
   // Do nothing.
 }
 
 //==============================================================================
 void SO3Joint::convertPositionsToState(
-    const Eigen::VectorXd& _positions, StateSpace::State* _state) const
+    const Eigen::VectorXd& positions, StateSpace::State* state) const
 {
   setQuaternion(
-      static_cast<State*>(_state),
-      SO3::Quaternion(BallJoint::convertToRotation(_positions)));
+      static_cast<State*>(state),
+      SO3::Quaternion(BallJoint::convertToRotation(positions)));
 }
 
 //==============================================================================
 void SO3Joint::convertStateToPositions(
-    const StateSpace::State* _state, Eigen::VectorXd& _positions) const
+    const StateSpace::State* state, Eigen::VectorXd& positions) const
 {
-  // TODO: We should call BallJoint::convertToRotation instead of logMap once
-  // the convertToRotation method is fixed in DART.
-  _positions = ::dart::math::logMap(
-      getQuaternion(static_cast<const State*>(_state)).toRotationMatrix());
+  positions = BallJoint::convertToPositions(
+      getQuaternion(static_cast<const State*>(state)).toRotationMatrix());
 }
 
 } // namespace dart